Refactoring of packages
* Split util into bricklist and output packages
* Split different repsonsibilities into different files
