+++ output/file.go
@@ -1,15 +1,13 @@
 /*
 Copyright © 2022 Oliver Götz <developer@geekgasm.eu>
 */
-package util
+package output
 
 import (
 	"bufio"
 
 	"fmt"
 	"os"
-
-	"golang.design/x/clipboard"
 )
 
 func WriteToFile(outputfile string, output string) {
@@ -31,14 +29,4 @@ func WriteToFile(outputfile string, output string) {
 	}
 	writer.Flush()
 	fmt.Printf("Output written to file %v\n", outputfile)
-}
-
-func CopyToClipboard(output string) {
-	err := clipboard.Init()
-	if err != nil {
-		fmt.Printf("Warning: clipboard could not be initialized, XML is not copied to clipboard.\n")
-		return
-	}
-	clipboard.Write(clipboard.FmtText, []byte(output))
-	fmt.Printf("Output copied to clipboard.\n")
 }
